[BEAM-2423] Abstract StateInternalsTest for the different state internals
